Add building_blocks hierarchy
