+++ hpccm/building_blocks/__init__.py
@@ -13,15 +13,3 @@
 # limitations under the License.
 
 from __future__ import absolute_import
-
-from hpccm.common import container_type
-from hpccm.common import linux_distro
-
-from hpccm.ConfigureMake import ConfigureMake
-from hpccm.Stage import Stage
-from hpccm.git import git
-from hpccm.recipe import recipe
-from hpccm.sed import sed
-from hpccm.tar import tar
-from hpccm.toolchain import toolchain
-from hpccm.wget import wget
