+++ hpccm/building_blocks/charm.py
@@ -24,7 +24,7 @@ from __future__ import print_function
 import logging # pylint: disable=unused-import
 import os
 
-from hpccm.packages import packages
+from hpccm.building_blocks.packages import packages
 from hpccm.primitives.comment import comment
 from hpccm.primitives.copy import copy
 from hpccm.primitives.environment import environment
