+++ hpccm/building_blocks/intel_psxe.py
@@ -25,7 +25,7 @@ import logging # pylint: disable=unused-import
 import os
 import re
 
-from hpccm.packages import packages
+from hpccm.building_blocks.packages import packages
 from hpccm.primitives.comment import comment
 from hpccm.primitives.copy import copy
 from hpccm.primitives.environment import environment
