+++ hpccm/building_blocks/mkl.py
@@ -27,15 +27,14 @@ import re
 
 import hpccm.config
 
+from hpccm.building_blocks.packages import packages
 from hpccm.common import linux_distro
-from hpccm.packages import packages
 from hpccm.primitives.comment import comment
 from hpccm.primitives.copy import copy
 from hpccm.primitives.environment import environment
 from hpccm.primitives.shell import shell
 from hpccm.toolchain import toolchain
 from hpccm.wget import wget
-from hpccm.yum import yum
 
 class mkl(wget):
     """MKL building block"""
