+++ hpccm/building_blocks/mlnx_ofed.py
@@ -25,8 +25,8 @@ import os
 
 import hpccm.config
 
+from hpccm.building_blocks.packages import packages
 from hpccm.common import linux_distro
-from hpccm.packages import packages
 from hpccm.primitives.comment import comment
 from hpccm.primitives.shell import shell
 from hpccm.tar import tar
