+++ hpccm/building_blocks/openmpi.py
@@ -15,7 +15,7 @@
 # pylint: disable=invalid-name, too-few-public-methods
 # pylint: disable=too-many-instance-attributes
 
-"""MVAPICH2 building block"""
+"""OpenMPI building block"""
 
 from __future__ import absolute_import
 from __future__ import unicode_literals
@@ -24,57 +24,55 @@ from __future__ import print_function
 import logging # pylint: disable=unused-import
 import os
 import re
-from copy import copy as _copy
 
 import hpccm.config
 
+from hpccm.building_blocks.packages import packages
 from hpccm.common import linux_distro
-from hpccm.packages import packages
 from hpccm.ConfigureMake import ConfigureMake
 from hpccm.primitives.comment import comment
 from hpccm.primitives.copy import copy
 from hpccm.primitives.environment import environment
 from hpccm.primitives.shell import shell
-from hpccm.sed import sed
 from hpccm.tar import tar
 from hpccm.toolchain import toolchain
 from hpccm.wget import wget
 
-class mvapich2(ConfigureMake, sed, tar, wget):
-    """MVAPICH2 building block"""
+class openmpi(ConfigureMake, tar, wget):
+    """OpenMPI building block"""
 
     def __init__(self, **kwargs):
         """Initialize building block"""
 
         # Trouble getting MRO with kwargs working correctly, so just call
         # the parent class constructors manually for now.
-        #super(mvapich2, self).__init__(**kwargs)
+        #super(openmpi, self).__init__(**kwargs)
         ConfigureMake.__init__(self, **kwargs)
-        sed.__init__(self, **kwargs)
         tar.__init__(self, **kwargs)
         wget.__init__(self, **kwargs)
 
-        self.__baseurl = kwargs.get('baseurl',
-                                    'http://mvapich.cse.ohio-state.edu/download/mvapich/mv2')
+        self.baseurl = kwargs.get('baseurl',
+                                  'https://www.open-mpi.org/software/ompi')
         self.__check = kwargs.get('check', False)
-        self.configure_opts = kwargs.get('configure_opts', ['--disable-mcast'])
+        self.configure_opts = kwargs.get('configure_opts',
+                                         ['--disable-getpwuid',
+                                          '--enable-orterun-prefix-by-default'])
         self.cuda = kwargs.get('cuda', True)
         self.directory = kwargs.get('directory', '')
-        self.__gpu_arch = kwargs.get('gpu_arch', None)
+        self.infiniband = kwargs.get('infiniband', True)
         self.__ospackages = kwargs.get('ospackages', [])
-        self.prefix = kwargs.get('prefix', '/usr/local/mvapich2')
+        self.prefix = kwargs.get('prefix', '/usr/local/openmpi')
         self.__runtime_ospackages = [] # Filled in by __distro()
-
-        # MVAPICH2 does not accept F90
-        self.toolchain_control = {'CC': True, 'CXX': True, 'F77': True,
-                                  'F90': False, 'FC': True}
-        self.version = kwargs.get('version', '2.3rc2')
-
-        self.__commands = []              # Filled in by __setup()
-        self.__environment_variables = {} # Filled in by __setup()
 
         # Input toolchain, i.e., what to use when building
         self.__toolchain = kwargs.get('toolchain', toolchain())
+        self.version = kwargs.get('version', '3.0.0')
+
+        self.__commands = [] # Filled in by __setup()
+        self.__environment_variables = {
+            'PATH': '{}:$PATH'.format(os.path.join(self.prefix, 'bin')),
+            'LD_LIBRARY_PATH':
+            '{}:$LD_LIBRARY_PATH'.format(os.path.join(self.prefix, 'lib'))}
         self.__wd = '/var/tmp' # working directory
 
         # Output toolchain
@@ -92,10 +90,10 @@ class mvapich2(ConfigureMake, sed, tar, wget):
 
         instructions = []
         if self.directory:
-            instructions.append(comment('MVAPICH2'))
+            instructions.append(comment('OpenMPI'))
         else:
             instructions.append(comment(
-                'MVAPICH2 version {}'.format(self.version)))
+                'OpenMPI version {}'.format(self.version)))
         instructions.append(packages(ospackages=self.__ospackages))
         if self.directory:
             # Use source from local build context
@@ -123,96 +121,63 @@ class mvapich2(ConfigureMake, sed, tar, wget):
 
         if hpccm.config.g_linux_distro == linux_distro.UBUNTU:
             if not self.__ospackages:
-                self.__ospackages = ['byacc', 'file', 'openssh-client', 'wget']
-            self.__runtime_ospackages = ['openssh-client']
+                self.__ospackages = ['bzip2', 'file', 'hwloc', 'make',
+                                     'openssh-client', 'perl', 'tar', 'wget']
+            self.__runtime_ospackages = ['hwloc', 'openssh-client']
         elif hpccm.config.g_linux_distro == linux_distro.CENTOS:
             if not self.__ospackages:
-                self.__ospackages = ['byacc', 'file', 'make',
-                                     'openssh-clients', 'wget']
-            self.__runtime_ospackages = ['openssh-clients']
+                self.__ospackages = ['bzip2', 'file', 'hwloc', 'make',
+                                     'openssh-clients', 'perl', 'tar', 'wget']
+            self.__runtime_ospackages = ['hwloc', 'openssh-clients']
         else: # pragma: no cover
             raise RuntimeError('Unknown Linux distribution')
 
-    def __set_gpu_arch(self, directory=None):
-        """Older versions of MVAPICH2 (2.3b and previous) were hard-coded to
-        use the "sm_20" GPU architecture.  Use the specified value
-        instead."""
-
-        if self.cuda and self.__gpu_arch and directory:
-            self.__commands.append(
-                self.sed_step(file=os.path.join(directory, 'Makefile.in'),
-                              patterns=[r's/-arch sm_20/-arch {}/g'.format(self.__gpu_arch)]))
-
     def __setup(self):
+
         """Construct the series of shell commands, i.e., fill in
            self.__commands"""
 
-        # Create a copy of the toolchain so that it can be modified
-        # without impacting the original.
-        toolchain = _copy(self.__toolchain)
-
-        tarball = 'mvapich2-{}.tar.gz'.format(self.version)
-        url = '{0}/{1}'.format(self.__baseurl, tarball)
+        # The download URL has the format contains vMAJOR.MINOR in the
+        # path and the tarball contains MAJOR.MINOR.REVISION, so pull
+        # apart the full version to get the MAJOR and MINOR components.
+        match = re.match(r'(?P<major>\d+)\.(?P<minor>\d+)', self.version)
+        major_minor = 'v{0}.{1}'.format(match.groupdict()['major'],
+                                        match.groupdict()['minor'])
+        tarball = 'openmpi-{}.tar.bz2'.format(self.version)
+        url = '{0}/{1}/downloads/{2}'.format(self.baseurl, major_minor,
+                                             tarball)
 
         # CUDA
         if self.cuda:
-            cuda_home = "/usr/local/cuda"
-            if toolchain.CUDA_HOME:
-                cuda_home = toolchain.CUDA_HOME
-
-            # The PGI compiler needs some special handling for CUDA.
-            # http://mvapich.cse.ohio-state.edu/static/media/mvapich/mvapich2-2.0-userguide.html#x1-120004.5
-            if toolchain.CC and re.match('.*pgcc', toolchain.CC):
+            if self.__toolchain.CUDA_HOME:
                 self.configure_opts.append(
-                    '--enable-cuda=basic --with-cuda={}'.format(cuda_home))
-
-                if not toolchain.CFLAGS:
-                    toolchain.CFLAGS = '-ta=tesla:nordc'
-
-                if not toolchain.CPPFLAGS:
-                    toolchain.CPPFLAGS = '-D__x86_64 -D__align__\(n\)=__attribute__\(\(aligned\(n\)\)\) -D__location__\(a\)=__annotate__\(a\) -DCUDARTAPI='
-
-                if not toolchain.LD_LIBRARY_PATH:
-                    toolchain.LD_LIBRARY_PATH = os.path.join(cuda_home,
-                                                             'lib64', 'stubs') + ':$LD_LIBRARY_PATH'
+                    '--with-cuda={}'.format(self.__toolchain.CUDA_HOME))
             else:
-                self.configure_opts.append(
-                    '--enable-cuda --with-cuda={}'.format(cuda_home))
-
-            # Workaround for using compiler wrappers in the build stage
-            self.__commands.append('ln -s {0} {1}'.format(
-                os.path.join(cuda_home, 'lib64', 'stubs', 'libnvidia-ml.so'),
-                os.path.join(cuda_home, 'lib64', 'stubs',
-                             'libnvidia-ml.so.1')))
-            self.__commands.append('ln -s {0} {1}'.format(
-                os.path.join(cuda_home, 'lib64', 'stubs', 'libcuda.so'),
-                os.path.join(cuda_home, 'lib64', 'stubs', 'libcuda.so.1')))
-
+                self.configure_opts.append('--with-cuda')
         else:
-            self.configure_opts.append('--disable-cuda')
+            self.configure_opts.append('--without-cuda')
+
+        # InfiniBand
+        if self.infiniband:
+            self.configure_opts.append('--with-verbs')
+        else:
+            self.configure_opts.append('--without-verbs')
 
         if self.directory:
             # Use source from local build context
-            self.__set_gpu_arch(
-                directory=os.path.join(self.__wd, self.directory))
             self.__commands.append(self.configure_step(
                 directory=os.path.join(self.__wd, self.directory),
-                toolchain=toolchain))
+                toolchain=self.__toolchain))
         else:
             # Download source from web
             self.__commands.append(self.download_step(url=url,
                                                       directory=self.__wd))
             self.__commands.append(self.untar_step(
                 tarball=os.path.join(self.__wd, tarball), directory=self.__wd))
-            self.__set_gpu_arch(
-                directory=os.path.join(self.__wd,
-                                       'mvapich2-{}'.format(self.version)))
-
             self.__commands.append(self.configure_step(
                 directory=os.path.join(self.__wd,
-                                       'mvapich2-{}'.format(self.version)),
-                toolchain=toolchain))
-
+                                       'openmpi-{}'.format(self.version)),
+                toolchain=self.__toolchain))
 
         self.__commands.append(self.build_step())
 
@@ -230,29 +195,15 @@ class mvapich2(ConfigureMake, sed, tar, wget):
             self.__commands.append(self.cleanup_step(
                 items=[os.path.join(self.__wd, tarball),
                        os.path.join(self.__wd,
-                                    'mvapich2-{}'.format(self.version))]))
-
-        # Setup environment variables
-        self.__environment_variables = {
-            'LD_LIBRARY_PATH':
-            '{}:$LD_LIBRARY_PATH'.format(os.path.join(self.prefix, 'lib')),
-            'PATH': '{}:$PATH'.format(os.path.join(self.prefix, 'bin'))}
-        if self.cuda:
-            # Workaround for using compiler wrappers in the build stage
-            self.__environment_variables['PROFILE_POSTLIB'] = '"-L{} -lnvidia-ml -lcuda"'.format('/usr/local/cuda/lib64/stubs')
+                                    'openmpi-{}'.format(self.version))]))
 
     def runtime(self, _from='0'):
         """Install the runtime from a full build in a previous stage"""
         instructions = []
-        instructions.append(comment('MVAPICH2'))
-        # TODO: move the definition of runtime ospackages
+        instructions.append(comment('OpenMPI'))
         instructions.append(packages(ospackages=self.__runtime_ospackages))
         instructions.append(copy(_from=_from, src=self.prefix,
                                  dest=self.prefix))
-        # No need to workaround compiler wrapper issue for the runtime.
-        # Copy the dictionary so not to modify the original.
-        vars = dict(self.__environment_variables)
-        if vars.get('PROFILE_POSTLIB'):
-            del vars['PROFILE_POSTLIB']
-        instructions.append(environment(variables=vars))
+        instructions.append(environment(
+            variables=self.__environment_variables))
         return instructions
