+++ hpccm/building_blocks/pgi.py
@@ -26,8 +26,8 @@ import os
 
 import hpccm.config
 
+from hpccm.building_blocks.packages import packages
 from hpccm.common import linux_distro
-from hpccm.packages import packages
 from hpccm.primitives.comment import comment
 from hpccm.primitives.copy import copy
 from hpccm.primitives.environment import environment
