+++ hpccm/building_blocks/python.py
@@ -23,7 +23,7 @@ from __future__ import print_function
 
 import logging # pylint: disable=unused-import
 
-from hpccm.packages import packages
+from hpccm.building_blocks.packages import packages
 from hpccm.primitives.comment import comment
 
 class python(object):
